Moving include in "wfc" subdirectory (renaming namespace)
